made activity checker (tradeactivitychecker.jsx) its own component. Called component inside of the politicianrecenttrade.jsx.
