+++ src/client/features/politicians/PoliticianRecentTrade.jsx
@@ -4,6 +4,7 @@ import { useParams } from "react-router-dom";
 import { useGetPoliticianQuery } from "./politicianSlice";
 import { useGetSenateDataQuery } from "./senateApiSlice";
 import { useGetHouseDataQuery } from "./houseApiSlice";
+import TradeActivityChecker from "./TradeActivityChecker";
 
 function PoliticianRecentTrade() {
   const { id } = useParams();
@@ -45,38 +46,6 @@ function PoliticianRecentTrade() {
   if (isError) {
     return <h1>Error loading data</h1>;
   }
-  const checkActivity = (politician, houseTrades, senateTrades) => {
-    const fullName = `${politician.first_name} ${politician.last_name}`;
-
-    if (politician.role === "Rep" && Array.isArray(houseTrades)) {
-      // Check if there are any transactions for the representative
-      const representativeTrades = houseTrades.filter((transaction) => {
-        return (
-          transaction.representative &&
-          transaction.representative
-            .toLowerCase()
-            .includes(fullName.toLowerCase())
-        );
-      });
-
-      return representativeTrades.length > 0;
-    } else if (politician.role === "Sen" && Array.isArray(senateTrades)) {
-      // Check if there are any transactions for the senator
-      const senatorTrades = senateTrades.filter((transaction) => {
-        return (
-          transaction.senator &&
-          transaction.senator.toLowerCase().includes(fullName.toLowerCase())
-        );
-      });
-
-      return senatorTrades.length > 0;
-    }
-
-    return false; // Default to false if role is not recognized
-  };
-
-  const activityStatus = checkActivity(politician, houseTrades, senateTrades);
-  console.log(`Account is ${activityStatus ? "active" : "inactive"}`);
 
   return (
     <section>
@@ -108,6 +77,13 @@ function PoliticianRecentTrade() {
             <p>No active trading.</p>
           )}
         </div>
+      )}
+      {politician.role && (
+        <TradeActivityChecker
+          politician={politician}
+          houseTrades={houseTrades}
+          senateTrades={senateTrades}
+        />
       )}
     </section>
   );
